修改对象元数据的大小属性为int类型

+++ common/src/main/java/com/jay/oss/common/entity/object/ObjectVO.java
@@ -21,7 +21,7 @@ import java.io.Serializable;
 public class ObjectVO implements Serializable {
     private String objectKey;
     private String versionId;
-    private Long size;
+    private Integer size;
     private Long createTime;
     private String md5;
     private String fileName;
